+++ lib/Hooks/UserHooks.php
@@ -24,7 +24,7 @@ class UserHooks {
 	public function register() {
 		$maintenance = $this->maintenance;
 		$callback = function ($user) use ($maintenance) {
-			$maintenance->resetDb($user->getUID());
+			$maintenance->resetAllData($user->getUID());
 		};
 		$this->userManager->listen('\OC\User', 'postDelete', $callback);
 	}