+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Small layout issues on Nextcloud 25
 - Layout issue in the two-line controls pane on IE
 - Not adjusting to dark theme when the theme comes from the browser preference (in NC25)
+- User's podcasts, radio stations, and Ampache/Subsonic API keys not erased when an user account deleted
 
 ## 1.7.0 - 2022-10-31
 ### Added